using unsigned int only

+++ inc/dstring_entry.h
@@ -9,7 +9,7 @@ struct dentry
 	dstring* key;
 
 	// cached hashvalue of the key
-	unsigned long long int key_hash_value;
+	unsigned int key_hash_value;
 
 	void* value;
 };
@@ -18,7 +18,7 @@ dentry* get_dentry(dstring* key, void* value);
 
 int key_compare_dentry(const void* entry1, const void* entry2);
 
-unsigned long long int key_hash_dentry(const void* entryp);
+unsigned int key_hash_dentry(const void* entryp);
 
 void delete_dentry(dentry* entryp, void (*value_destroyer)(void* value));
